mbedtls: Activate deterministic ECDSA
With deterministic ECDSA the value k needed for the ECDSA signature is
not randomly generated any more, but generated from a hash over the
private key and the message to sign. If the value k used in a ECDSA
signature or the relationship between the two values k used in two
different ECDSA signatures over the same content is know to an attacker
he can derive the private key pretty easily. Using deterministic ECDSA
as defined in the RFC6979 removes this problem by deriving the value k
deterministically from the private key and the content which gets
signed.

The resulting signature is still compatible to signatures generated not
deterministic.

This increases the size of the ipk on mips 24Kc by about 2 KByte.
old:
163835 libmbedtls_2.9.0-1_mips_24kc.ipk
new:
165712 libmbedtls_2.9.0-1_mips_24kc.ipk

This does not change the ECDSA performance in a measurable way.

Source: https://git.openwrt.org/?p=openwrt/staging/hauke.git;a=commitdiff;h=87a47466eb0172bb8f929a0a54df423884b9cf37
Author: Hauke Mehrtens <hauke@hauke-m.de>

Modified the commit/patch to apply and follow code style

Signed-off-by: Daniel Engberg <daniel.engberg.lists@pyret.net>